♻️ Refactor: 방 관련 타입코드 통합 및 .message 추가
방 관련 타입코드 통합 및 .message 추가
