--- src/styles/fonts.css
@@ -1,15 +1,0 @@
-@font-face {
-  font-family: 'Pretendard';
-  font-weight: 600;
-  src: url('/fonts/Pretendard-SemiBold.woff') format('woff');
-}
-@font-face {
-  font-family: 'Pretendard';
-  font-weight: 500;
-  src: url('/fonts/Pretendard-Medium.woff') format('woff');
-}
-@font-face {
-  font-family: 'Pretendard';
-  font-weight: 400;
-  src: url('/fonts/Pretendard-Regular.woff') format('woff');
-}
