+++ src/types/create.ts
@@ -1,5 +1,4 @@
-/* 모캠프(방) 타입 정의 */
-
+// 방 생성 Form 데이터
 export interface CreateRoomFormData {
   roomName: string;
   capacity: number;
